// demo/src/sqlite/sqlite.d.ts
export interface ImportObject {
  'sqlite:wasm/extension-callbacks@0.1.0': typeof SqliteWasmExtensionCallbacks,
  'sqlite:wasm/extension@0.1.0': typeof SqliteWasmExtension,
  'wasi:cli/exit@0.2.0': typeof WasiCliExit,
  'wasi:cli/stderr@0.2.0': typeof WasiCliStderr,
  'wasi:cli/stdin@0.2.0': typeof WasiCliStdin,
  'wasi:cli/stdout@0.2.0': typeof WasiCliStdout,
  'wasi:cli/terminal-input@0.2.0': typeof WasiCliTerminalInput,
  'wasi:cli/terminal-output@0.2.0': typeof WasiCliTerminalOutput,
  'wasi:cli/terminal-stderr@0.2.0': typeof WasiCliTerminalStderr,
  'wasi:cli/terminal-stdin@0.2.0': typeof WasiCliTerminalStdin,
  'wasi:cli/terminal-stdout@0.2.0': typeof WasiCliTerminalStdout,
  'wasi:clocks/monotonic-clock@0.2.0': typeof WasiClocksMonotonicClock,
  'wasi:clocks/wall-clock@0.2.0': typeof WasiClocksWallClock,
  'wasi:filesystem/preopens@0.2.0': typeof WasiFilesystemPreopens,
  'wasi:filesystem/types@0.2.0': typeof WasiFilesystemTypes,
  'wasi:io/error@0.2.0': typeof WasiIoError,
  'wasi:io/poll@0.2.0': typeof WasiIoPoll,
  'wasi:io/streams@0.2.0': typeof WasiIoStreams,
  'wasi:sockets/tcp@0.2.0': typeof WasiSocketsTcp,
  'wasi:sockets/udp@0.2.0': typeof WasiSocketsUdp,
}
export interface Root {
  'sqlite:wasm/low-level@0.1.0': typeof SqliteWasmLowLevel,
  lowLevel: typeof SqliteWasmLowLevel,
  'sqlite:wasm/high-level@0.1.0': typeof SqliteWasmHighLevel,
  highLevel: typeof SqliteWasmHighLevel,
  'sqlite:wasm/extension@0.1.0': typeof SqliteWasmExtension,
  extension: typeof SqliteWasmExtension,
}